# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Bubble Tea Shop API",
    "description" : "Spring-owned HTTP contract for guest ordering and staff operations.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/guest/orders" : {
      "post" : {
        "tags" : [ "Guest ordering" ],
        "summary" : "Place an idempotent pending cash order",
        "operationId" : "placeGuestOrder",
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateGuestOrderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Existing order replayed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuestOrder"
                }
              }
            }
          },
          "201" : {
            "description" : "Order placed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuestOrder"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/Problem"
          },
          "401" : {
            "$ref" : "#/components/responses/Problem"
          },
          "403" : {
            "$ref" : "#/components/responses/Problem"
          },
          "409" : {
            "$ref" : "#/components/responses/Problem"
          },
          "503" : {
            "$ref" : "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/guest/locations/{locationSlug}/orders" : {
      "post" : {
        "tags" : [ "Guest ordering" ],
        "summary" : "Place an order at a public location",
        "operationId" : "placeGuestLocationOrder",
        "parameters" : [ {
          "name" : "locationSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateGuestOrderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Existing order replayed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuestOrder"
                }
              }
            }
          },
          "201" : {
            "description" : "Order placed",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuestOrder"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/Problem"
          },
          "409" : {
            "$ref" : "#/components/responses/Problem"
          },
          "503" : {
            "$ref" : "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "properties" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
...
      "CreateGuestOrderLineRequest" : {
        "type" : "object",
        "properties" : {
          "variantId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1
          },
          "optionChoiceIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "maxItems" : 30,
            "minItems" : 0
          }
        },
        "required" : [ "optionChoiceIds", "variantId" ]
      },
      "CreateGuestOrderRequest" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CreateGuestOrderLineRequest"
            },
            "maxItems" : 25,
            "minItems" : 0
          }
        },
        "required" : [ "items" ]
      },
      "GuestOrder" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "publicOrderNumber" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string"
          },
          "currencyCode" : {
            "type" : "string"
          },
          "subtotalMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "replayed" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GuestOrderLine"
            }
          }
        }
      },
      "GuestOrderLine" : {
        "type" : "object",
        "properties" : {
          "productName" : {
            "type" : "string"
          },
          "variantName" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unitPriceMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lineTotalMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "options" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GuestOrderOption"
            }
          }
        }
      },
      "GuestOrderOption" : {
        "type" : "object",
        "properties" : {
          "groupName" : {
            "type" : "string"
          },
          "choiceName" : {
            "type" : "string"
          },
          "priceDeltaMinor" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    },
    "responses" : {
      "Problem" : {
        "description" : "Problem details",
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ProblemDetail"
            }
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}